requirements.txt
Babel==2.3.4
colorama==0.3.7
decorator==4.0.10
Django==1.10.3
django-phonenumber-field==1.1.0
googlemaps==2.4.4
phonenumberslite==7.7.4
pickleshare==0.7.4
prompt-toolkit==1.0.9
Pygments==2.1.3
pytz==2016.7
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
win-unicode-console==0.5
